# paginate array only for multi-word searches